requirements.txt
easy_thumbnails
django-tinymce
